Editando codigo para caso de estouro de inteiro "INT >> LONG"

+++ Mini-MaxSum.cs
@@ -1,13 +1,13 @@
  public static void miniMaxSum(List<int> arr)
  {
    arr.Sort();
-   int totalSum = 0;
+   long totalSum = 0;
    foreach (int num in arr)
    {
      totalSum += num;
    }
-   int minSum = totalSum - arr[arr.Count - 1];
-   int maxSum = totalSum - arr[0];
+   long minSum = totalSum - arr[arr.Count - 1];
+   long maxSum = totalSum - arr[0];
    Console.WriteLine(minSum + " " + maxSum);
  }
